Clean up About and Help page syntax
- Remove parent elements of `Typography`
The `Typography` element replaces itself with an HTML element that makes sense for the given `type`. For example, `type="body1"` is replaced with a `<p>` element, making it redundant to surround `<Typography type="body1">` in a `<p>` element.
- Change `Typography` `body2` to `body1`
`body2` renders to an `<aside>` element, which is not semantically correct for the type of text that those elements contained. `body1` renders to a `<p>` element, which makes more sense.
- Fix indentation, spacing, and line length
- Move `Typography` element to outside (rather than inside) of `<ul>`
It is more correct to wrap the entire `<ul>` rather than having a `<p` element containing the `<li>` elements inside the `<ul>`. `<li>` elements should be direct children of `<ul>`.
- Remove unused import from `NavLinks`