Fix routing for multiple matching paths
When one or more routes have the same base path (such as `/activities` and `/activities/:sessionCode/:activityId`), React Router will render both routes whenever they are both valid. This can be avoided by using a `<Switch>` around the routes, which tells the router to render the routes *exclusively*. This technique, when employed in `App` and `Header`, fixes the problem of multiple routes being selected at the same time. The only caveat is that routes must be defined in order of most specific to least specific. Ambiguity in routes must be avoided when possible (for example, `/activities/:activityId` and `/activities/:type` are ambiguous because the router can't tell whether `activityId` or `type` is being provided).