Fix global menu offsetting header
The use of `#root` to set the app containers to `height: 100%` was changed because Sass lint does not allow the use of IDs as selectors. However, using `body > div` has the side effect of applying the `height` property to the `div` used for the global menu, too, which caused the header offset problem. `#root` with a comment that disables Sass lint for that line fixes the problem.