Add automatic code formatting with Prettier
- Add Prettier dependency
- Replace Sasslint with Stylelint
Stylelint has a "recommended" preset that will only look for syntax errors, which is desirable because it should not check code style now that Prettier is in use.
- Replace AirBnB ESLint config with create-react-app config
The AirBnB config is opinionated and checks style, which is no longer necessary now that Prettier is in use. The create-react-app config only checks for syntax errors. Also, it is unsupported to use a custom ESLint config when using create-react-app, so this change brings the project more in line with best practices.
- Add a guide on code style
- Add Prettier and Stylelint extensions to recommended VS Code extensions list
- Enable VS Code "format on save" setting for Prettier extension
- Move errant development dependencies to correct section of `package.json`
- Add `lint-staged` to format and lint only staged files (instead of all files in the project) in the pre-commit script
- Remove comments for disabling Sasslint and ESLint on specific lines