Fixed apartment people search
The search bar will now correctly pass the username of the selected student via the callback `onSearchSubmit`, and the application page will retrieve the StudentProfileInfo object from the API and store it in a list. The ApplicantListField component will then create the textfields to display the students in the list.
The styling needs to be fixed, but the functionality appears to work as intended