Add dialog to reduce false positives, refactor
This commit adds a pre-submit confirmation dialog when the user is reporting positive for COVID 19 symptoms.
The intent is to ensure that only people who mean to report symptoms will do so, since currently many people
click yes and submit without realizing that that is reporting symptoms.

Additionally, this commit includes a good deal of refactoring. With the addition of the dialog, the Question component had ballooned in complexity.
The refactoring extracts some of that (i.e. API response formatting) to the wellness service and organizes the component further for clarity.