Add network context
This commit introduces the NetworkContext, which uses React's Context API to track the online/offline status of the user's network. The goal is to have network-aware component's subscribe to this context provider so that the network is tracked in only one place, rather than having eventListeners in every component that needs to know the network status.

This may result in minimal performace gains (having only one event listener attached to the window instead of >10 at once), but I would guess that the rerendering takes significantly more processing than the eventlisteners themselves. The real advantage to this technique is the separation of concerns. Now, the NetworkContextProvider is the only component responsible for the logic of tracking network changes. Every ther network-aware component simply subscribes to these changes with the very minmial hook useNetworkStatus - e.g. `const networkStatus = useNetworkStatus`, rather than having a large block of logic duplicated across many components with a useEffect hook and window listeners that obscures the component's more natural responsibilities.