Refactor right side navbuttons
This commit refactors the right side navbuttons. The main purpose was to make use of the useNetworkIsOnline hook. In the process, I also cleaned up the GordonRightNavButtons significantly and streamlined the navbuttons to reuse a single JSX component instead of repeating small variations of the same logic and styling once for each link.