Simplify event service
The event service was fetching events in two steps - using a plain fetch function and then a formatting and sorting function that called the plain fetch function. The plain fetch function was exposed beyond the service  but was never used (and would never be used since the unformatted events can't be displayed well).