Create GitHub Actions CI/CD
Since Travis-CI is no longer a viable solution, this commit switches to
GitHub Actions by implementing a CI/CD workflow. The workflow, in ci/yml
lints and builds a branch whenever a new commit is pushed to it, and
also uploads build artifacts from the develop and master branches for
deployment.