Fix timesheets rounding bug.
Use a non-strict comparison operator, >, to compare the roundedHourDifference to our minimum and round it up where necessary. This ensures that, despite the type mismatch that is necessary to force hours worked to display as a two digit number (which can only be achieved via string in JavaScript), the comparison is still accurate.