Fix CLW credits
CLW Credits required default was broken because it depended on CLWAttendedEvents, so we couldn't fetch the actual value until the student had attended some event in the current session. The API was fixed to add ChapelRequired and ChapelAttended to the StudentProfileInfo. This commit uses those values for the CLWCreditsDaysLeft wheel.