Fix date formatting for timezone bug
Sending dates as ISO strings to the backend caused an unforeseen problem in that dates were being parsed by the backend as UTC and stored in the database as such, but then compared to local time without adjusting for the timezone difference. Instead, we will send the date in a custom format that approximates ISO but is in the user's local time zone. This fixes the above error and preserves the fix from using toLocaleString that caused issues for user's with browsers that had different locales.