Refactor hook separate from NetworkContext
The NetworkContext can still be used in Class Components via the Context.Consumer component in the React Context API, so to support our legacy class components, we will make the NetworkContext available that way as well as via the hook. Prefer the hook for it's clarity and performance wherever feasible.