Replacing favicons containing a white bg with those having a transparent bg
I made a copy of the old favicons under old-images that can be deleted once we're ready to merge. Could I get feedback on whether this was the intended solution to the issue @EjPlatzer @cpabbot ? By the way, the new favicons are used on my laptop on firefox and phone on firefox, but I don't see it changing for my iphone safari even though I cleared the site data for the local network host.