Refactor Header component
This refactor required significant changes because the custom hook useNetworkStatus can only be run in functional components. The refactoring is desirable, since functional components are React best practices, and they have greatly reduced the component's complexity.

Note, for example, the >250 line (50%) reduction in component length. Lines of code do not equate directly with complexity, but the metric is a useful approximation. In this case, it demonstrates the reduction of repated code (using one function for all tabs, rather than having nearly the same code block repeated many times, hiding the meaningful differences), the removal of Class Component boilerplate, and the extraction of the network status logic.