Update or uninstall vulnerable dependencies
We had >11 high severity vulnerabilities in our dependency tree. These have now been resolved.

React-scripts was updated to version 4, which did not result in any breaking changes for our project. The old version depended on a version of http-proxy that had vulnerabilities.

The unused testing dependencies (Jasmine, Karma, etc.) were removed because they are currently totally unused and karma had vulneable dependencies.

The url-regex package had a vulnerability which didn't affect our site but was polluting the results of npm audit. That package is no longer maintained, but url-regex-safe is an update that fixes the vulnerability and improves the package's API.

http-proxy-middleware was updated because it's old version depended on a version of http-proxy which had high severity vulnerabilities. The new http-proxy-middleware has a slightly changed API that requires us to call it slightly differently. setupProxy.js has been adjusted accordingly.

react was updated to fit the rest of the dependencies properly. The major version 17 has virtually no breaking changes and no new features. It is mostly an internal update preparing for future changes in coming major versions.

date-fns and @date-io/date-fns were updated as well.