Fix svg imports and clipPath ids
Older versions of react-scripts (the create-react-app package) had our bundler configured to allow relative imports of svgs for display. However, the new version of react-scripts no longer supports this. Instead, we use SVGR to import svgs as React components. This is overall a nicer practice, but it did reveal a problem with our victory-promise svgs. The `clipPath` tag in each svg had the same `id` attribute ('clip-path'), which means they were using the value of the first one to be displayed. The result was that each svg was only drawing to the width of the first svg (which was also the narrowest).