Fix & Restore Timesheets after temporary outage
Timesheets was taken down temporarily because of an issue where shifts were being submitted with no Time information. It turns out that there were options on Intl.DateTimeFormat JavaScript built-in we were using (specifically `dateStyle` and `timeStyle`) that aren't supported by Safari. This was causing safari users to lose the time of the their shift start and end DateTimes.