Fix timesheets reload when type in End Time field
Timesheets had an annoying bug where the view would reload itself whenever someone typed in the End Time field. This is because no validation was being done on dates before sending them to the backend to fetch jobs. So the user would start typing an End Time, but the first couple of letters of their End Time would not be a valid date, so we would send an invalid which would cause an Error that would force the page to reload in production.