Simplify PersonalInfoList on Profile & MyProfile
The component formerly called ProfileList listed personal information of users. It was used in both Profile and MyProfile.

It had a great deal of redundant code, and that repeated code was obscuring the actual logic of the component.
This made maintaining and adding to the component very difficult, which is an issue because the component handles personal info that is sensitive in nature. Although the API ensures info cannot be leaked to people who aren't authorized to view it, this component handled displaying what parts of the information shown are considered sensitive and when.