Group shared Profile/MyProfile components
The OfficeInfoList, PersonalInfoList, Identification, and MembershipsList components were all only used in Profile and MyProfile. Having them all in `src/components` was cluttering the folder, making it harder to find the more generally shared components. Grouping them in the ProfileInfo folder decreases the clutter and also opens the opportunity to import them in a more unified way.