Refactor and update InvolvementsAll
The InvolvementsAll (formerly ActivitiesAll) view was in desperate need of refactoring. I have switched the view and all it's components over to functional components using hooks. I have made use of the useNetworkStatus hook. I have also streamlined the design and made it more consistent with other pages (including consistent use of cards, fixing the stickiness of the filters, fixing and streamlining the layout of the requestsReceived component, etc.).