First Commit
Modified all the news-related front-end files/ The main task was adding all the photo submission functionality into the news js file, because the existing code I could look at was not easily converted into its class component-based equivalent.

No compile errors as of right now, and no visible place to add an image to the news post- that will be the next task.