Small updates
Working on backend, so these changes are just removing some now-unnecessary logging. Did finally refactor the two handle submission methods. Was hoping to keep them discrete so that converting this to a functional component in the future might be easier, but it was having issues that got solved by cleaning up a little, so that's in the past.