Fix UserContext issues with initialValue
Now UserContext will white out the screen while we load user info for the first time. If the user is not logged-in, the extra white screen is not noticeably longer than previous loading times. If the user IS authenticated, it takes a tiny bit longer, but it should end up making the site feel snappier overall, once the redundant fetches of profile info are removed. Also, since this only happens on initial site load (i.e. when the user first opens the site in a tab), it is an acceptable trade-off.