Refactor UserContext for better cohesion
UserContext's first implementation was somewhat naive, consisting of simply a `user` object which, when null, represented the unauthenticated state, and when valued was a `profile` object. This limits the possibilities of UserContext substantially.

Now, UserContext uses a reducer, which means it can handle more complex state updates that affect multiple parts of the state object at once. This also allows us to separate out parts of the user data to making accessing them simpler. The new version will allow for more flexibility and the possibility to add more to the UserContext in the future.

The new structure includes a separate `AuthContext` that only tracks the authentication state, accessible via the `useAuth` hook. This allows components to subscribe to just the auth state, so that they don't have to re-render every time part of the user data changes.

The new structure also allows for destructuring the `user` returned by `useUser`, becuase it will never be null (since auth state is tracked separately and the reducer guarantees that the object's shape is preserved in certain ways). For example, a component can destructure (and thereby subscribe to) JUST the `images` portion of user data now via `const { images } = useUser()`. This will be especially helpful if we add more, less frequently accessed data to the user object.