Fix auth error on loading site
Within the `src/components/Profile/components/PersonalInfoList/index.js` file, there was a call to `userService.getLocalInfo()` that was outside any component. That meant it was compiled as part of the static JS rather than part of any React component, and so it was getting run when the site loaded. This is a problem because it would mean the value never updated (even if the logged-in user changed), and also because it was run even if the user wasn't authenticated. The latter case would cause `user.js` to throw an auth error because there's no token to read the user's local info from. Because the code was run outside of React, the error was thrown before React came online, meaning it prevented the entire site from loading at all.

I have changed it so that the relevant code is now inside the `PersonalInfoList` component. The value it fetches from local info is memoized, to prevent unnecessary work. Since `PersonalInfoList` is only ever rendered to an authenticated user, it won't cause an auth error (and if it did, it would now be caught by the error boundary). The memoization with no dependencies is fine because in order for the value to change, the user would have to log out and back in, so the component would unmount and remount.