Add nvmrc file to tell nvm to use LTS version
The Node Version Manager can be run as `nvm install` and `nvm use` when
a `.nvmrc` file is present. This configuration file specifies which
version of Node NVM should install/use. Since we want all users to be on
the LTS (Long-Term Support) version, including a .nvmrc file fits our
usecase for nvm well.