Reformat package-lock.json for npm >=7.x
As of npm 7.x, the `package-lock.json` file has changed formats.
The previous format is now `lockfileVersion: 1`, and npm >=7.x uses
`lockfileVersion: 2`. This is mostly a syntactic change, and only needs
to happen once. As long as all devs are using the LTS version of Node
(currently 16), npm will be >7.x