Fetch Slider images from DB again
We have restructured the slider images. The image metadata is now stored in the CCT database (rather than the now obsolete WWW database). The images themselves are stored in the api filesystem. Currently, the image files and metadata have to be manually managed by ISG, but it should be simple enough to create a page for site admins to manage them.