WIP: Use MSAL/Azure AD for authentication
With the API upgrade, we would've needed to rebuild our authentication structure nearly from scratch. The old authentication strategy used LDAP to our local Active Directory domain, and it required manually implementing token creation and reading in the API. REcreating this in ASP.NET Core would've been a large undertaking. In the meantime, Gordon has begun updating other applications to use Azure Active Directory (the cloud-based auth strategy). Implementing Azure AD Auth in the API was much simpler and quicker than rebuilding the old local LDAP structure. It also comes with many benefits - the foremost being a simple way to used AD Groups an authorization mechanism.

However, using Azure AD authentication affected the front end somewhat extensively, requiring a deep rewrite in some places to effectively use the new authentication scheme. Most of these changes will be a net positive in terms of security, developer experience, and user experience, over the old system. This is a preliminary implementation of the new authentication scheme in the front end, using the Microsoft Authentication Library (MSAL). The new scheme consists of using MSAL to authenticate to Azure AD, receiving an ID token and access token from Azure AD after authenticating, and using that access token to authenticate to and authorize with the API.

There is still much polishing and fixing to be done with this new scheme, but this is a mostly working prototype.