Load profile gracefully
Rather than showing just a spinner on a blank page while we're loading the user, we would like to let them continue to interact with the website until their profile info is loaded.

the msal-react library provides a useIsAuthenticated hook, which tells us whether we have a valid authentication token from Azure AD (i.e. whether we can make requests to the backend). Previously, the UserContextProvider included a context for the useAuth hook, which would indicate whether the user was authenticated AND their profile was loaded. This was redundant with simply checking whether the profile was not null.

Thus, I have reviewed every use of isAuthenticated to determine whether it needs to wait for the user's profile to be loaded or whether it only cares about the authentication token. In the former case, I got rid of isAuthenticated and just checked whether profile is loaded directly. In the latter case, I updated isAuthenticated to read from useIsAuthenticated (from the msal-react library).

What this means is, we no longer flash a blank screen/just a spinner while the user's profile is loading after login. Instead, they login and see the site as it normally is. Anyting that depends on the profile will wait until it's loaded before updating, whereas anything that depends on just the authentication token will immediately recognize the new authentication state and begin sending requests to the backend using the token.

The end result is a smoother experience, and slightly less time spent waiting for API requests to finish, because more requests can happen in parallel while we're still waiting on the user's profile.