Add user profile loading status and show loader
Now that auth status and user profile have been decoupled, there were odd cases where the user would sign in and see a message like "You are not logged in. Please log in to use this page" in the time between setting their authentication status and loading their profile information from the API.

I have added a `loading` property to the User object returned from `useUser`. This `loading` property is a boolean, which is true if the user is authenticated but their profile information is still loading, and false otherwise. This can be used to show a loader (or something more sophisticated) in the case that the user has signed in but we don't yet have their information which is needed to load content.