Reimplement search time in quick search
Although quick search is currently fast enough (and debounced) that it is rare that a earlier query would return later then a subsequent query, it is still not impossible. So I re-implemented the search time guard - searching now returns a tuple of the time when the search was started and the search results. If the time the search was started is greater than the time the the current results were saved, then the new search results replace the current results (and the time is updated).

In order to get this working smoothly without a rendering loop, I stored searchtime and search results in a tuple. That way the updater function can compare new and current search time without maintaining a reference to the current search time state.