Set apiBaseURL based on environment
In development (i.e. when running the front-end locally via `npm run start`), we want API requests to be sent to the front-end node server, and re-directed by the http-proxy-middleware to the appropriate API. This is because the proxy sets the `origin` header of API requests to the API url, thereby circumventing CORS errors while developing 360 locally.

While we could prevent CORS errors by allowing cross-origin requests from localhost in the API, that would be a bad idea. Essentially, that would allow nefarious third-party websites or applications on a user's machine to impersonate us and send requests on their behalf, compromising our API using their credentials.

Instead, this solution allows us to connect to any of our API servers while running the API locally, but ensures that from end user's browsers, only 360[train] is allowed to send requests to the API.

When not developing locally (i.e. when running a static build deployed to one of our web servers), we neither need nor have a proxy server to overwrite request's `origin` header. So, we send requests directly to the API, knowing that each API is configured to accept cross-origin requests from it's own front-end (i.e. 360apitrain allows cross-origin requests from the 360train origin).