Extract components from People Search
The People Search page component was more than 1000 lines long and dealt with several things at once. It was confusing and difficutl to maintain. In this commit, I have extracted the search fields into a separate component, encapsulating that portion of the page to make the rest of the page more straight forward.

I have also stream-lined and simplified the search fields, and extracted the most repetitive field logic into smalled components. There are some edge cases that I haven't figured out yet, notably class year because it's the only case where the display is different from the value. I can also extract and imporove the CSS, which will clarify the code greatly.