Replace react-text-mask with react-imask
react-text-mask has been unmaintained since 2018, and was officially shutdown by the previous maintainer in December of 2020. It was causing dependency resolution issues because it depends on react <=16, but most of our other dependencies have upgraded to v17 (and will soon update to v18).

IMask.js is a popular cross-framework implementation of the same functionality. It is well-maintained and is a recommended replacement for react-text-mask.