Use PersonType to filter profile info temporarily
We can't use AuthGroups to decide what profile info is relevant for now, becuase we can only access AuthGroups of the authenticated user. This will likely change in the future, but for now, we will revert to using PersonType to determine what info is relevant.

Note that this is not really related to security, since the API controls what profile info is sent. This just ensures that we're only trying to show the correct kind of info on profiles of a given type.