People Search: Sync search params to/from URL
This enables the "deep-linking", aka URLs that point to a particular state of the People Search page rather than the default state. Any search params stored in the query string of the URL will update the search inputs of the page. Crucially, this only happens when the page is first loaded, i.e. when someone navigates to the page either via a link, the URL bar, or the back/forward buttons. Relatedly, pressing the search button will - in addition to searching - push a new history entry with the current search params in the URL query string (only if they are different from the current URL query string). Consequently, users can go back and forth between their searches - and also share links to a particular search.