Inject MUI styles first for easier overrides
And fix weird styles on Office Info List.

By wrapping our app in a MUI StylesProvider with the `injectFirst` attribute, we tell Material-UI to inject it's stylesheet into the page BEFORE any other CSS is added. This means that, by default, other CSS has a higher priority than MUI simply by virtue of coming after it. This is in contrast to MUI's default behavior of injecting its stylesheet at the very end of the `head` of the page, which makes it override any other CSS of equal specificity.