People Search: Improve search param handling
I documented the `searchParamSerializerFactory` and the functions it creates more thoroughly. Since there is a significant amount of complexity, this documentation should help future maintainers (including myself) understand, update, and use it well.

I also fixed some bugs in the serializtion/deserialization of search params. Now, the history will not be affected when performing a search with the same parameters as the current query string. Also, when resetting search params (and loading defaults), we now handle privileges for students and alumni more consistently.