MUI Upgrade: Apply codemod preset-safe
This codemod provided by the MUI team performs some non-breaking code changes. The most notable is updating import paths, since the path of all MUI assets has changed from v4 to v5.

Other updates including setting some props on elements to preserve the old default, when there is a new default value in v5. For example, `IconButton` has a prop `size`, which defaulted to `large` in v4 but now defaults to a slightly smaller size in v5. So, the codemod added the prop `size="large"` to any IconButtons which didn't previously have an explicit size set.