Encapsulate membership interval logic
The transcript displays the interval(s) that a user was a member of an involvement. An interval is an unbroken sequence of memberships, where the user was a member for every (significant) session.

The logic for transforming a list of session codes into a string representing the intervals of a user's membership is relatively complex. Previously, it was done inside the `formatDuration` function and several helper functions. These functions interweaved complex mutation, long comments, and a lot of magic string handling. This resulted in a confusing control flow and a complex algorithm spread across multiple error-prone pieces. I have encapsulated most of the logic into a new MembershipInterval class, and simplified `formatDuration`.