Extract MSAL Config to fix circular dependency
The Microsoft Authentication Library (MSAL) package we use for Auth requires a single instance of the PublicClientApplication to be created at runtime. This instance requires some complex configuration, which was kept in `src/services/auth.ts`, but the instance has to be created once at runtime when the app starts up, so it was being done in `src/index.tsx`. The problem is that all the other methods in `auth.ts` require a reference to the MSAL instance, so `auth.ts` was importing the instantiated object from `index.tsx`. This created a circular dependency where index.tsx imports from auth.ts and then auth.ts imports from index.tsx.

The solution is to move the MSAL instance configuration out of auth.ts and into the component where the initialization actually happens. That way, index.tsx does not import auth.ts at all, resolving the circular dependency. Becuase that configuration logic is large and not especially relevant to the app initialization that is done in index.tsx, I extracted it (and the rest of the Context Provider logic) into separate components.