Fix PeopleSearch params not updating on navigation
For posterity - somehow, the `popstate` event listener was not being triggered within `SearchFieldList` when the user navigated in history. The fix was to add the event listener within a separate useEffect that only runs on mount, rather than inside the one that runs everytime `location` from React-Router updated.

Somehow, this bug in the event listener didn't appear until the Header component called `useLocation`. We have no idea why that is the case. :shrug: