Read params from URL directly in event listener
Previously, we listened for the `popstate` event and updated a `ref` when the
event happened. Simultaneously, we had a useEffect that would run whenever
`location` changed, which would update the search params from the URL *if* the
`ref` said to. This *almost* worked, except that, on the very first `popstate`
event, the ref would not update before the `useEffect` ran, meaning we wouldn't
read the params from the URL the first time that the user triggered `popstate`
after `SearchFieldList` mounted.

The solution is to simply perform the update (i.e. read the params from the URL)
directly inside the event listener callback. The only times we need to read the
params from the URL are when the component first mounts (in case we're mounting
with a paramful URL, i.e. because the user opened 360 from a deeplink), and when
the `popstate` event is triggered. So there's really no need to have this
indirection of a ref and a callback watching `location`, since we don't depend
on the `location` but are simply reacting to the event.