Remove unnecessary useCallback dependency
The `search` callback depended on `location.search` from react-router's `useLocation` hook. However, that hook is intended to be used to force React to ..."react" to changes to the location, i.e. to cause React to re-render when the location changes. That is not necessary in the case of this callback, we don't need the component to re-render, we just need the callback to reference the current value of `location`. The much simpler soultion is to use `window.location.search` from the browser. Since the global `window` object is managed by the browser, it is always kept up to date, so the value in the callback will never be stale. This also has the advantage of not forcing the component to re-render and update the callback everytime the URL changes.