Fix quick search z index
Apparently `position: sticky` creates a new stacking context with a default z-index of 0. This caused the Qucik search results in the header to appear underneath any raised content on the page, because it could never go higher than it's parent's stacking context z-index of 0. Setting the app bar to have a z-index of 100 resolves the issue.