Code-split top-level routes for smaller js files
Using the dynamic `import()` function that imports a JS Module only when called, we can split our code into several discrete chuncks of code, as opposed to having it all in one giant file. This allows us to ensure that users only have to download the code needed for the pages they view, instead of downloading the code for the entire site even if they only visit the homepage.

This problem is technically caused by our bundler, which transforms the thousands of files of source code into unified bundles of code in one file. Bundling is desirable because it prevents the browser from having to "waterfall" request the source code: i.e. request App.js from the server, and then request Home.js that is imported by App.js, and then request CLWCredits.js that is imported by Home.js, etc. The only downside is that the bundled file can become so large that it takes a user-noticeable amount of time to download, and includes code that will never run because the user only visits a small portion of the site. Code-splitting lets us optimize in both directions, by bundling our code into files exactly sized to display one top-level route of the site each. So the user gets only the code they need for the page they are currently viewing, but they get all that code in one request that serves one reasonably-sized file.